# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Gorax - a powerful workflow automation platform with webhooks, scheduling, and real-time execution monitoring.\n\n## Authentication\nAll API endpoints (except /health, /ready, and webhook triggers) require authentication.\nIn development mode, use the X-User-ID header. In production, use Ory Kratos session cookies.\n\n## Multi-tenancy\nAll authenticated endpoints require a valid X-Tenant-ID header to identify the tenant context.\n\n## Rate Limiting\nAPI requests are rate-limited based on tenant quotas. Check response headers for limit information.\n\n## Error Handling\nErrors follow a consistent format with appropriate HTTP status codes and error messages.",
        "title": "Gorax Workflow Automation API",
        "contact": {
            "name": "Gorax Support",
            "url": "https://github.com/gorax/gorax",
            "email": "support@gorax.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/workflows/bulk/clone": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Clones multiple workflows in a single operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Bulk clone workflows",
                "parameters": [
                    {
                        "description": "Bulk clone request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BulkOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cloned workflows and result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/bulk/delete": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Deletes multiple workflows in a single operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Bulk delete workflows",
                "parameters": [
                    {
                        "description": "Bulk delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BulkOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_workflow.BulkOperationResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/bulk/disable": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Disables multiple workflows in a single operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Bulk disable workflows",
                "parameters": [
                    {
                        "description": "Bulk disable request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BulkOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_workflow.BulkOperationResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/bulk/enable": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Enables multiple workflows in a single operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Bulk enable workflows",
                "parameters": [
                    {
                        "description": "Bulk enable request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BulkOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gorax_gorax_internal_workflow.BulkOperationResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/workflows/bulk/export": {
            "post": {
                "security": [
                    {
                        "TenantID": []
                    },
                    {
                        "UserID": []
                    }
                ],
                "description": "Exports multiple workflows as JSON in a single operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Workflows"
                ],
                "summary": "Bulk export workflows",
                "parameters": [
                    {
                        "description": "Bulk export request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BulkOperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export data and result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gorax_gorax_internal_workflow.BulkOperationFailure": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "workflow_id": {
                    "type": "string"
                }
            }
        },
        "github_com_gorax_gorax_internal_workflow.BulkOperationResult": {
            "type": "object",
            "properties": {
                "failures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gorax_gorax_internal_workflow.BulkOperationFailure"
                    }
                },
                "success_count": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.BulkOperationRequest": {
            "type": "object",
            "required": [
                "workflow_ids"
            ],
            "properties": {
                "workflow_ids": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "Ory Kratos session cookie (production mode)",
            "type": "apiKey",
            "name": "ory_kratos_session",
            "in": "cookie"
        },
        "TenantID": {
            "description": "Tenant identifier for multi-tenant isolation",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        },
        "UserID": {
            "description": "User identifier (development mode only)",
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    },
...
}